requirements.txt
pandas
piexif
argparse
flask
tqdm
mysql-connector-python-rf
psycopg2-binary>=2.9.0
python-dotenv 